Reconnect when stuffing data into the stream failed
This should never happen when everything is operatin normally, because
we only try to send data if the stream signalls it has space.
If the write fails, the underlying TCP socket is in unexpected state, a
direct reconnect without waiting for a timeout should help to reduce the
time until we are connected again.
This can save about 2-3 seconds on app open.